+++ src/components/Game/Header/UserHeader.tsx
@@ -1,14 +1,21 @@
 import React, {FunctionComponent} from 'react';
 import styles from './UserHeader.module.scss'
 import avatar from '../../../assets/png/avatar.png'
+import TheTimer from "../../Timer/TheTimer";
 
-const UserHeader: FunctionComponent = () =>
+type UserHeaderProps = {
+    username: string
+}
+
+const UserHeader: FunctionComponent<UserHeaderProps> = ({username}) =>
     <div className={styles.UserHeader}>
         <div className={styles.avatar}>
-            <img src={avatar}/>
+            <TheTimer>
+                <img src={avatar}/>
+            </TheTimer>
         </div>
         <div className={styles.user}>
-            <div>JujuOneLove</div>
+            <div>{username}</div>
             <div>1er</div>
         </div>
     </div>