+++ src/hooks/useInterval.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useEffect, useRef } from 'react';
 
 function useInterval(callback:()=>void, delay:number) {
   const savedCallback:any = useRef();
@@ -14,8 +14,8 @@ function useInterval(callback:()=>void, delay:number) {
       savedCallback.current();
     }
     if (delay !== null) {
-      let id = setInterval(tick, delay);
-      return () => clearInterval(id);
+      const interval = setInterval(tick, delay);
+      return () => clearInterval(interval);
     }
   }, [delay]);
 }