+++ src/mixins/shuffle.ts
@@ -1,11 +1,14 @@
 /**
- * Shuffles array in place. ES6 version
+ * Shuffles array in place.
  * @param {Array} a items An array containing the items.
  */
-export function shuffle(a:[]) {
-    for (let i = a.length - 1; i > 0; i--) {
-        const j = Math.floor(Math.random() * (i + 1));
-        [a[i], a[j]] = [a[j], a[i]];
+export function shuffle(a:any[]) {
+    let j, x, i;
+    for (i = a.length - 1; i > 0; i--) {
+        j = Math.floor(Math.random() * (i + 1));
+        x = a[i];
+        a[i] = a[j];
+        a[j] = x;
     }
     return a;
 }