+++ src/pages/Labs.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
+import TheTimer from '../components/Timer/TheTimer'
 import {Link} from "react-router-dom";
-import Timer from '../components/Timer/Timer'
+import Cards from '../components/Cards/Cards'
 
 const Labs = () => {
     return (
         <div className="Labs">
             <h1>LABS</h1>
-            <Timer />
+            <TheTimer />
+            <Cards />
         </div>
     );
 }