+++ src/server/actions/createRoom.ts
@@ -10,7 +10,14 @@ export function createRoom(rooms: Rooms) {
     let room: Room = {
         id: nanoid(5),
         sockets: [],
-        users: []
+        users: [],
+        colors: ['red', 'blue', 'yellow', 'green', 'purple', 'orange'],
+        game: {
+            points: 100,
+            round: 0,
+            idUser: 0,
+            isStart: false
+        }
     }
     while (rooms[room.id] || !room?.id) {
         room.id = nanoid(5);