+++ src/server/index.ts
@@ -6,6 +6,12 @@ import { joinRoom } from "./actions/joinRoom";
 import {createRoom} from "./actions/createRoom";
 import {leaveRooms} from "./actions/leaveRooms";
 import {ExtendedSocket} from "./types/socket";
+import axios from "axios";
+import {initIdOMG} from "./actions/initOMG";
+import {shuffle} from "../mixins/shuffle";
+import {getPlayer} from "./actions/getPlayer";
+import {nextRound} from "./actions/nextRound";
+import {socket} from "../socketClient";
 
 const app = express();
 const server = require('http').createServer(app);
@@ -37,7 +43,8 @@ io.on("connect", (socket: ExtendedSocket) => {
     socket.on('joinRoom', (username:string, roomId:string, callback) => {
         const room:Room = rooms[roomId];
         if(room) {
-            if(room.sockets.length < 6 || room.users.length < 6) joinRoom(username, socket, room);
+            if(room.game.isStart) socket.emit('error', "La room est en plein jeu");
+            else if(room.sockets.length < 6 || room.users.length < 6) joinRoom(username, socket, room);
             else socket.emit('error', "La room est complète");
         }
         else socket.emit('error', "La room n'existe pas");
@@ -53,6 +60,69 @@ io.on("connect", (socket: ExtendedSocket) => {
             socket.emit('updateUsers', room.users);
         }
     });
+
+    /**
+     * Gets fired for get my color in room
+     */
+    socket.on('getMyColor', () => {
+        socket.emit('getMyColor', socket.color);
+    });
+
+    /**
+     * Gets fired when a host start a game in room.
+     */
+    socket.on('startGame', async (roomId: string) => {
+        const room:Room = rooms[roomId];
+
+        //Initialize cards for the game
+        await axios.get('https://happiness-strapi.herokuapp.com/cards').then(({data}) => {
+            room.game.cards = shuffle(data)
+        })
+
+        //Initialize guesses for the game
+        await axios.get('https://happiness-strapi.herokuapp.com/guesses').then(({data}) => {
+            room.game.guesses = shuffle(data)
+        })
+
+        if(room.game.guesses && room.game.guesses.length > 0) room.game.idGuesses = 0
+
+        //Initialize OMG for the game
+        room.game.idOMG = initIdOMG(room)
+
+        room.game.isStart = true
+        socket.emit('redirect', `/game/${roomId}`);
+        socket.to(room.id).emit('redirect', `/game/${roomId}`);
+    });
+
+    socket.on('nextRound', (roomId: string) => {
+        const room:Room = rooms[roomId];
+
+        nextRound(room, socket)
+    })
+
+    socket.on('endOMG', (roomId: string) => {
+        const room:Room = rooms[roomId];
+
+        //Re-initialize OMG for the game
+        room.game.idOMG = initIdOMG(room)
+
+        socket.emit('endOMG')
+
+        nextRound(room, socket)
+    });
+
+    /**
+     * Get fired for get player in game room
+     */
+    socket.on('getPlayer', (roomId: string) => {
+        const room:Room = rooms[roomId];
+
+        getPlayer(room, socket)
+    });
+
+    socket.on('gameOver', (roomId: string) => {
+        socket.to(roomId).emit('redirect', `/end`);
+    })
 
     /**
      * Gets fired when a player leaves a room.