+++ src/server/types/rooms.ts
@@ -1,10 +1,13 @@
 import {Socket} from "socket.io";
-import {User} from "./users";
+import {User} from "../../store/user/types";
+import {Game} from "./game";
 
 export interface Room {
     id: string,
     sockets: Socket[],
-    users: User[]
+    users: User[],
+    colors: string[],
+    game: Game
 }
 
 export interface Rooms {