+++ src/store/user/action.ts
@@ -1,4 +1,11 @@
-import {SET_NAME, UserActionTypes} from "./types";
+import {SET_ID, SET_NAME, UserActionTypes} from "./types";
+
+export function setId(id: string): UserActionTypes {
+  return {
+    type: SET_ID,
+    payload: id
+  }
+}
 
 export function setName(newName: string): UserActionTypes {
   return {