+++ src/store/user/reducer.ts
@@ -1,11 +1,12 @@
 import {Reducer} from "redux";
 
-import {SET_NAME, UserActionTypes, userState} from "./types";
+import {SET_ID, SET_NAME, UserActionTypes, userState} from "./types";
 
 export const initialState: userState = {
   data: {
+    id: '',
     name: '',
-    jwt: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NCwiaWF0IjoxNjEzMDcwODI4LCJleHAiOjE2MTU2NjI4Mjh9.podMQsLLLWnnRRPdbViwUyaF2jZlh5SkTYxiVIljxH4'
+    color: undefined
   },
   errors: undefined,
   loading: false
@@ -13,6 +14,9 @@ export const initialState: userState = {
 
 const reducer: Reducer<userState, UserActionTypes> = (state:userState = initialState, action: UserActionTypes) => {
   switch (action.type) {
+    case SET_ID: {
+      return { ...state, data: {...state.data, id: action.payload,} };
+    }
     case SET_NAME: {
       return { ...state, data: {...state.data, name: action.payload,} };
     }